Lighting changes added to forked repo for PR consideration
Pulled in changes to the mushr_cad database on branch 'add-lighting-system'  to support the addition of a lighting system to the MuSHR robotic vehicle platform based on Adafruit NeoPixel RGBW LED Sticks and Trinket-M0 MCU for control. Match documented changes in separate repo [pfliegster/mushr-lighting-system](https://github.com/pfliegster/mushr-lighting-system), except for change #3 which is pending.